+++ src/mlpack/methods/huber_regression/huber_regression_impl.hpp
@@ -17,33 +17,13 @@
 
 namespace mlpack {
 
-template<typename MatType, typename VecType>
-HuberLoss<MatType, VecType>::HuberLoss(const MatType& X, const VecType& y, const double delta)
-    : X(X), y(y), delta(delta) {}
+
 
 template<typename MatType, typename VecType>
-double HuberLoss<MatType, VecType>::Evaluate(const MatType& theta) {
-  VecType residuals = y - X * theta;
-  VecType huber = arma::zeros<VecType>(residuals.n_elem);
-  for (size_t i = 0; i < residuals.n_elem; ++i) {
-    if (std::abs(residuals[i]) <= delta) {
-      huber[i] = 0.5 * std::pow(residuals[i], 2);
-    } else {
-      huber[i] = delta * (std::abs(residuals[i]) - 0.5 * delta);
+HuberRegressor<MatType, VecType>::HuberRegressor(const MatType& X, const VecType& y,double delta, size_t maxIter, double tol)
+    : delta(delta), maxIter(maxIter), tol(tol) {
+      Train(X,y);
     }
-  }
-  return arma::accu(huber);
-}
-
-template<typename MatType, typename VecType>
-void HuberLoss<MatType, VecType>::Gradient(const MatType& theta, MatType& gradient) {
-  VecType residuals = y - X * theta;
-  gradient = -X.t() * (residuals % arma::clamp(residuals / delta, -1, 1));
-}
-
-template<typename MatType, typename VecType>
-HuberRegressor<MatType, VecType>::HuberRegressor(double epsilon, size_t maxIter, double tol)
-    : epsilon(epsilon), maxIter(maxIter), tol(tol) {}
 
 template<typename MatType, typename VecType>
 void HuberRegressor<MatType, VecType>::Train(const MatType& X, const VecType& y) {
@@ -51,7 +31,7 @@ void HuberRegressor<MatType, VecType>::Train(const MatType& X, const VecType& y)
   coef = VecType(X.n_cols, arma::fill::zeros);
 
   // Define the Huber loss function
-  HuberLoss<MatType, VecType> loss(X, y, epsilon);
+  HuberLoss<MatType, VecType> loss(X, y, delta);
 
   // Define the L-BFGS optimizer
   ens::L_BFGS optimizer(maxIter, X.n_cols * 10, tol);
